Convert the target image to float32 for speed.
I don't know how this got skipped. Convert to float32 along with all the
rest.

+++ test.py
@@ -38,7 +38,7 @@ def parse_args():
 
 
 def main(args):
-    target = imageio.imread(args.image)
+    target = imageio.imread(args.image).astype("float32")
     approximation = np.zeros_like(target, dtype="float32")
 
     if args.ea:
